write article under subdir named after article category
ideally we would like to have one category per wiki article. that's
not always the case. and sometimes the article has no category or has
a category not matching those we set in settings.toml — which build up
the navigation. in both cases, we use a fallback category (see
settings.toml).