build-article: make has_duplicates func
this check if same article filepath is in different cat
sub-dir. useful check when the category of an article has changed —
given we have no easy way to check for that from the MW API (afaik),
this fs check seems way easier to use and should be fine perf wise too.