build-wiki: remove --metadata-only
this simplified the code a lot. since we're always getting back the
HTML of the wiki article, there's no advantage in keeping the double
code to just fetch and parse the 'metadata' of the article, since
we're not parsing the raw wikitext into HTML anymore.

this has also another advantage: we can use the pre-rendered article
HTML to extract a list of image URLs, instead of doign another HTTP
request for each image in an article to get the actual URL of it (this
is due to the fact that otherwise the MW APIs returns us only a list
of image filenames, not URLs).