update-categories: basic implementation working
what's missing is checking via article's backlinks when another
article is updated. for eg:

- the current article (a) i'm looking at is mentioned (linked) in another
article (b) , and it is listed in the backlinks
- this other article (b) is updated meanwhile

right now the current article (a) won't be updated as well in the
category pages nor in the article page itself, in the case that the
other article (b) has removed a link to it (to article `a`).

this should clearly take place though. a first way would be to use the
list of backlinks that each article has (if any), and update each
article both as a single page as well as part of an cat-index page.