update-category-index: prepare event article as needed
move out most of the logic in make_event_index to own func so that we
can re-use it in other places. this new function mostly parse and
prepare event's date and such.

so far this is the only article category that does more prep work
before being sent to the template, so it's a one-off if branch check.