article delete + restore: update cat-page and rebuild backlink pages
when delete an article, we remove any link or snippet present in any
other article pointing to the delete article.

when restoring a deleted article, we parse every link in the restored
article that is a wiki page and re-build each page so that it displays
the missing backlink to the restored page. plus we rebuild the
category page to which the restored article belongs to.