make-redirect-article: fix bug from Path.stem when a dot in a string
eg, for some leftover reasons, we were converting the article_title
into a Path object and do a Path(article_title).stem to remove the
filename extension (i believe?). if there was a dot in the
articl_title, this Path().stem operation would keep everything before
the dot, causing to truncate the article_title. no match would be find
in the next line etc.

eg `This title. has a dot` => `This title`.

simplify code as we pass an article_title, not a filepath-like string.